// src/Types.tsx
export type EventType = {
  name: string;
  location: string;
  availability: boolean;
  time: string;
  day: string;
  description: string;
  _id: string;
};

export type AddEventType = {
  name: string;
  location: string;
  availability: boolean;
  time: string;
  day: string;
  description: string;
};